avgRate and roomDto

+++ src/main/java/com/example/hotelk/review/domain/response/ReviewResponse.java
@@ -15,7 +15,7 @@ import java.sql.Timestamp;
 public class ReviewResponse {
     private Hotel hotelName;
     private String guestName;
-    private Integer rating;
+    private Double rating;
     private String comment;
 
     public ReviewResponse(Review review) {
